+++ packages/shared/widget-plugin-filtering/src/typings/PickerFilterStore.ts
@@ -1,4 +1,4 @@
-import { RefFilterStore } from "../stores/RefFilterStore";
+import { RefFilterStore } from "../stores/picker/RefFilterStore";
 import { StaticSelectFilterStore } from "../stores/picker/StaticSelectFilterStore";
 
 export type PickerFilterStore = RefFilterStore | StaticSelectFilterStore;
